Move runner files to application folder
