+++ applications/flash_attention_v2/flash_decode_runner.hpp
@@ -51,7 +51,7 @@
 #include "cutlass/util/reference/device/tensor_compare.h"
 
 #include "cutlass/../../examples/common/sycl_common.hpp"
-#include "cutlass/../../fmha_decode_configuration.hpp"
+#include "flash_attention_v2/fmha_decode_configuration.hpp"
 
 using namespace cute;
