+++ benchmarks/flash_attention/flash_attention_decode/benchmarks.hpp
@@ -32,7 +32,7 @@
 #pragma once
 
 #include "benchmark_runner.hpp"
-#include "cutlass/../../fmha_decode_configuration.hpp"
+#include "flash_attention_v2/fmha_decode_configuration.hpp"
 #include "benchmarks_h64_512.hpp"
 #include "benchmarks_h64_1024.hpp"
 #include "benchmarks_h96_512.hpp"
