+++ examples/sycl/06_bmg_flash_attention/bmg_flash_attn_decode_runner.hpp
@@ -30,7 +30,7 @@
  **************************************************************************************************/
 #pragma once
 
-#include "cutlass/../../flash_decode_runner.hpp"
+#include "flash_attention_v2/flash_decode_runner.hpp"
 #include "helper.h"
 #include "sycl_common.hpp"
